refactor(desktop): extract pane layout persistence
